+++ PIPS-IPM/Core/QpStoch/sVars.C
@@ -76,10 +76,10 @@ sVars::sVars( sTree* tree, OoqpVector * x_in, OoqpVector * s_in,
 	      OoqpVector * w_in, OoqpVector * phi_in,
 	      OoqpVector * t_in, OoqpVector * lambda_in,
 	      OoqpVector * u_in, OoqpVector * pi_in,
-	      OoqpVector * ixlow_in, int nxlowGlobal,
-	      OoqpVector * ixupp_in, int nxuppGlobal,
-	      OoqpVector * iclow_in, int mclowGlobal,
-	      OoqpVector * icupp_in, int mcuppGlobal)
+	      OoqpVector * ixlow_in, long long nxlowGlobal,
+	      OoqpVector * ixupp_in, long long nxuppGlobal,
+	      OoqpVector * iclow_in, long long mclowGlobal,
+	      OoqpVector * icupp_in, long long mcuppGlobal)
   : QpGenVars()
 {
